% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analog_novelty.R
\name{analog_novelty}
\alias{analog_novelty}
\title{Novelty (outlyingness) measurement for pre-determined spatial climate analogs.}
\usage{
analog_novelty(
  clim.targets,
  clim.analogs,
  label.targets,
  label.analogs,
  vars,
  clim.icvs = NULL,
  label.icvs = NULL,
  weight.icv = 0.5,
  sigma = TRUE,
  analog.focal = NULL,
  threshold = 0.95,
  pcs = NULL,
  plotScree = FALSE,
  plot2d = FALSE,
  plot2d.pcs = cbind(c(1, 2, 3, 4), c(2, 3, 4, 5)),
  plot3d = FALSE,
  plot3d.pcs = c(1, 2, 3),
  biplot = TRUE,
  plot3d.candidates = FALSE
)
}
\arguments{
\item{clim.targets}{data.table. Climate variables for which the analogs were
identified}

\item{clim.analogs}{data.table. Climate variables of at least 50 locations
representing the spatial variation in the reference period (historical) climate
of each analog in the analog pool.}

\item{label.targets}{character. Vector of the analog IDs identified for the
climatic conditions listed in \code{clim.targets}. Length equals number of records
in \code{clim.targets}.}

\item{label.analogs}{character. Vector of the analog IDs for the climatic
conditions listed in \code{clim.analogs}. Length equals number of records in
\code{clim.analogs}.}

\item{vars}{character. Climate variables to use in the novelty measurement.}

\item{clim.icvs}{data.table. Time series of climate variables at the geographic
centroids of each analog in the analog pool. If not null, this interannual climatic
variability will be pooled with the spatial variation of the analog to calculate
the covariance matrix used in the Mahalanobis distance measurement.}

\item{label.icvs}{character. Vector of the analog IDs for the climatic
conditions listed in \code{clim.icvs}. Length equals number of records in \code{clim.icvs}.}

\item{analog.focal}{character. Optionally specify a single analog for visualization.}

\item{threshold}{numeric. The cumulative variance explained to use as a threshold
for truncation of principal components to use in the mahalanobis distance measurement.}

\item{pcs}{integer. The fixed number of PCs to use in the mahalanobis distance
measurement. Non-null values of this parameter override the \code{threshold} parameter.}

\item{plotScree}{logical. If \code{analog.focal} is specified, plot a scree plot showing
the standard deviation of the analog points, target points, and (if activated) interannual
climatic variability (ICV). The difference of the analog and target means is also shown.}

\item{plot2d}{logical. If \code{analog.focal} is specified, plot a 4-panel set of
bivariate scatterplots visualizing the distances in the principal components.}

\item{plot2d.pcs}{numeric matrix. a 4x2 matrix indicating the principal components to
display in the 2D scatterplots. Rows correspond to the panel to be displayed,
and columns indicate the principal components to display in the x and y axes.}

\item{plot3d}{logical. If \code{analog.focal} is specified, plot a 3-dimensional scatterplot
visualizing the distances in the first three PCs.}

\item{plot3d.pcs}{numeric. Principal components to display in the x, y, and z
axes of the 3D scatterplot.}

\item{biplot}{logical. Include lines on the 3D plot indicating the correlation
of the standardized climate variables with the principal components.}

\item{plot3d.candidates}{logical. Include representative points and labelled centroids for all
western North American candidate analogs.}
}
\value{
\code{vector} of sigma dissimilarity (if sigma==TRUE) or Mahalanobis distances
(if sigma==FALSE) corresponding to each element of the 'analogs.target' vector.
}
\description{
This function calculates novelty (or outlyingness) of climate conditions by measuring
the Mahalanobis distance between these target climates and the centroid of their pre-determined
climate analogs. The covariance matrix of the Mahalanobis distance is the spatial
and/or interannual variation in climate conditions of the climate analog, as
selected by the user. It optionally renders 2D and 3D plots for visualizing
the distance measurements in the principal component space.
}
\details{
The novelty measure is calculated using Mahalanobis distance, which quantifies the
statistical distance between the target climate conditions and their spatial analogs
(i.e., climates similar to the target in historical records). The function can also
incorporate interannual climatic variability (ICV) in the covariance matrix of the climate
analog to provide additional information on the climatic scale of each principal component.
For visualization purposes, the function supports 2D and 3D scatterplots
based on principal component analysis (PCA), showing how the target climate conditions
relate to the analogs and, optionally, to the ICV.

The main output of the function is a vector of novelty values, either as Mahalanobis
distances or as sigma dissimilarities (Interpretable as the number of standard deviations away from
the analog cluster). These values are returned for each target climate in the dataset.
Additional outputs include several optional plots:
\itemize{
\item A scree plot showing the standard deviation of the first few principal components
\item A 2D scatterplot (if \code{plot2d} is enabled) showing the relationship between the target
and analogs across selected principal components
\item A 3D scatterplot (if \code{plot3d} is enabled) to visualize the distribution of the
target and analogs in 3D PCA space, with optional biplot vectors representing the
correlation of the standardized climate variables with the principal components.
}

The function supports customization of the principal component selection (via \code{threshold} or
\code{pcs}), weighting of the ICV in the distance measurement (\code{weight.icv}), and the selection
of specific analogs for visualization (\code{analog.focal}).
}
\examples{
if (FALSE) {
  
}
#'
}
