% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analog_novelty.R
\name{analog_novelty_core}
\alias{analog_novelty_core}
\title{Fast basic novelty function.}
\usage{
analog_novelty_core(
  clim.targets,
  clim.analogs,
  label.targets,
  label.analogs,
  vars,
  clim.icvs = NULL,
  label.icvs = NULL,
  weight.icv = 0.5,
  sigma = TRUE,
  threshold = 0.95,
  pcs = NULL
)
}
\arguments{
\item{clim.targets}{data.table. Climate values for which the analogs were
identified. Must include variables specified in \code{vars}.}

\item{clim.analogs}{data.table. Climate values of at least 50 locations
representing the spatial variation in the reference period (historical) climate
of each analog in the analog pool. Must include variables specified in \code{vars}.}

\item{label.targets}{character. Vector of the analog IDs identified for the
climatic conditions listed in \code{clim.targets}. Length equals number of records
in \code{clim.targets}.}

\item{label.analogs}{character. Vector of the analog IDs for the climatic
conditions listed in \code{clim.analogs}. Length equals number of records in
\code{clim.analogs}.}

\item{vars}{character. Climate variables to use in the novelty measurement.}

\item{clim.icvs}{data.table. Time series of climate variables at the geographic
centroids of each analog in the analog pool. If not null, this interannual climatic
variability will be pooled with the spatial variation of the analog to calculate
the covariance matrix used in the Mahalanobis distance measurement.}

\item{label.icvs}{character. Vector of the analog IDs for the climatic
conditions listed in \code{clim.icvs}. Length equals number of records in \code{clim.icvs}.}

\item{threshold}{numeric. The cumulative variance explained to use as a threshold
for truncation of principal components to use in the mahalanobis distance measurement.}

\item{pcs}{integer. The fixed number of PCs to use in the mahalanobis distance
measurement. Non-null values of this parameter override the \code{threshold} parameter.}

\item{clim.point}{data.table. Climate values of a single location of interest.
Must include variables specified in \code{vars}.}

\item{analog.focal}{character. Optionally specify a single analog for visualization.}
}
\value{
\code{vector} of sigma dissimilarity (if sigma==TRUE) or Mahalanobis distances
(if sigma==FALSE) corresponding to each element of the 'analogs.target' vector.
}
\description{
This function uses the same algorithm as \code{analog_novelty}, but has been paired down to remove plotting options
and is approx 15 times faster \code{analog_novelty}. We recommend using \code{analog_novelty} for most cases, unless the
data you're calculating novelty for are very large.
}
\examples{
if (FALSE) {
  
}
#'
}
