% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_analysis.R
\name{bgc_map}
\alias{bgc_map}
\title{Create a map of biogeoclimatic units with the standard BGC colour scheme}
\usage{
bgc_map(
  X,
  dat,
  zone = TRUE,
  boundary = NULL,
  mask = NULL,
  legend = FALSE,
  mask_alpine = TRUE,
  mask_ocean = TRUE,
  label_exotic = NULL,
  q_exotic = 0.5,
  title = "",
  add = FALSE
)
}
\arguments{
\item{X}{SpatRaster. A template raster.}

\item{dat}{data.table with first column indicating template raster cell ID and second column the BGC subzone/variant.}

\item{zone}{logical. Plot with BGC zone colours.}

\item{boundary}{SpatVector. Optional study area boundary.}

\item{mask}{SpatVector. Optional spatial mask}

\item{legend}{logical. Plot a legend.}

\item{mask_alpine}{logical. Plot alpine BGC units as white.}

\item{label_exotic}{numeric. Minimum number of grid cells required to plot a label for exotic (non-BC) BGC units.}

\item{q_exotic}{numeric. Quantile of cell numbers for positioning the label for exotic BGC units. Acceptable values are between 0 and 1.}

\item{title}{character. Title for the plot, left justified at top.}

\item{add}{logical. Add the map to an existing plot.}
}
\description{
This is a utility function that plots a basic map
}
