% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_analysis.R, R/cciss_duckdb.R
\name{bgc_persist_expand}
\alias{bgc_persist_expand}
\title{Create table of BGC subzone/variant persistance/expansion from raw BGC projections}
\usage{
bgc_persist_expand(dbCon, by_zone = TRUE)

bgc_persist_expand(dbCon, by_zone = TRUE)
}
\arguments{
\item{bgc_mapped}{List containing SpatRaster of BGCs and id table, or data.table (must have columns \code{cell},\code{BGC}). Usually created using \code{make_bgc_template}}

\item{periods}{Character. Either "auto" or vector of period names. If "auto", all periods present in bgc projections will be used.}

\item{base_folder}{Base folder to read results from.}
}
\value{
data.table containing persistance and expansion for each model/period/scenario/bgc

data.table containing persistance and expansion for each model/period/scenario/bgc
}
\description{
Create table of BGC subzone/variant persistance/expansion from raw BGC projections

Create table of BGC subzone/variant persistance/expansion from raw BGC projections
}
