% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_duckdb.R
\name{bgc_persist_expand_region}
\alias{bgc_persist_expand_region}
\title{Calculate BGC persistance and expansion by region}
\usage{
bgc_persist_expand_region(
  con,
  region_table,
  region_name,
  by_zone = TRUE,
  table_name = "bgc_per_exp_region"
)
}
\arguments{
\item{con}{duckdb connection}

\item{region_table}{character. Name of table with regions and corresponding cellnum}

\item{region_name}{character. Name of column in region table to group by}

\item{by_zone}{logica. Summarise by zone or subzone/variant? Default \code{TRUE}.}

\item{table_name}{table name to create/append to in database. Default is "bgc_per_exp_region"}
}
\value{
NULL. Writes table to database.
}
\description{
Calculate BGC persistance and expansion by region
}
