% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BGCs_notin_THLB}
\alias{BGCs_notin_THLB}
\title{Biogeoclimatic subzone-variants that are generally considered to be outside the Timber Harvesting Land Base}
\format{
\subsection{\code{BGCs_notin_THLB}}{

A data table with 211 rows and 2 columns:
\describe{
\item{BGC}{Biogeoclimatic subzone-variant}
\item{Exclude}{Excluded from THLB}
}
}
}
\usage{
BGCs_notin_THLB
}
\description{
Biogeoclimatic subzone-variants that are generally considered to be outside the Timber Harvesting Land Base
}
\keyword{datasets}
