% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_spatial.R
\name{cciss_full}
\alias{cciss_full}
\title{Internal function to create cciss projections from site series predictions. Calculates projected suitability for each period, as well as other CCISS statistics.}
\usage{
cciss_full(SSPred, suit, spp_select)
}
\arguments{
\item{SSPred}{Data.table of site series predictions. Usually comes from \code{siteseries_preds}}

\item{suit}{Data.table. Environmental suitability table}

\item{spp_select}{Character vector of species to use.}
}
\value{
data.table containing cciss results by siteno, period, edatope and species
}
\description{
Internal function to create cciss projections from site series predictions. Calculates projected suitability for each period, as well as other CCISS statistics.
}
