% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_duckdb.R
\name{cciss_full_species}
\alias{cciss_full_species}
\title{Calculate CCISS results using duckdb (single species)}
\usage{
cciss_full_species(con, spp, table_name = "cciss_res")
}
\arguments{
\item{con}{duckdb connection}

\item{spp}{character. Single species code to calculate for.}

\item{table_name}{table name to create/append to in database. Default is "cciss_res"}
}
\value{
NULL. Writes table to database.
}
\description{
Calculate CCISS results using duckdb (single species)
}
\details{
This function uses the same logic as \code{cciss_full}, but is orders of magnitude faster as it does all computation in duckdb, with all periods and edatopes at once. Eventually we will replace \code{cciss_full} with this function.
}
