% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_spatial.R
\name{cciss_rasterize}
\alias{cciss_rasterize}
\title{Create geotif rasters of projected suitabilities for each species/edatope/period}
\usage{
cciss_rasterize(raster_template, base_folder = "spatial")
}
\arguments{
\item{raster_template}{Template SpatRaster. Must be the same raster used throughout process}

\item{base_folder}{Base folder to write results to.}
}
\value{
NULL. Writes result geotifs to base_folder/final_rasters
}
\description{
Create geotif rasters of projected suitabilities for each species/edatope/period
}
