% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_spatial.R
\name{cciss_suitability}
\alias{cciss_suitability}
\title{Create projected suitability values from site series predictions.}
\usage{
cciss_suitability(dbCon, species, obs = FALSE, tile_size = 4000)
}
\arguments{
\item{dbCon}{duckdb database connection}

\item{species}{Character vector. Species codes to create projections for.}

\item{tile_size}{Integer. Number of sites to process at once. May need to decrease if memory is limited. Default 4000}
}
\value{
NULL. Writes table to database
}
\description{
Create projected suitability values from site series predictions.
}
