% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_duckdb.R
\name{dbAddDistricts}
\alias{dbAddDistricts}
\title{Add natural resource districts to database}
\usage{
dbAddDistricts(dbCon, raster_template, districts)
}
\arguments{
\item{dbCon}{duckdb database connection}

\item{raster_template}{raster used to populate database}

\item{districts}{sf object containing district boundaries (we suggest using \code{bcmaps::nr_districts()})}
}
\value{
NULL. Writes table to database
}
\description{
Add natural resource districts to database
}
