% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_duckdb.R
\name{dbAddRegions}
\alias{dbAddRegions}
\title{Add region table to duckdb for grouping}
\usage{
dbAddRegions(dbCon, raster_template, regions, region_name, table_name)
}
\arguments{
\item{dbCon}{duckdb database connection}

\item{raster_template}{raster used to populate database}

\item{regions}{SpatVect object containing district boundaries (we suggest using \code{bcmaps::nr_districts()})}

\item{region_name}{Name of column in \code{regions} containing names/codes}

\item{table_name}{character. Name of table to populate in database}
}
\value{
NULL. Writes table to database
}
\description{
Add region table to duckdb for grouping
}
