% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_duckdb.R
\name{dbCon_cciss}
\alias{dbCon_cciss}
\title{Connect to (and possibly create) DuckDB for storing results}
\usage{
dbCon_cciss(db_file, read_only = FALSE, threads = NULL, verbose = TRUE)
}
\arguments{
\item{db_file}{Character. Name of database file, or path. Must end in .duckdb}

\item{read_only}{Logical. Default FALSE}

\item{threads}{Numeric. Number of threads to use in database.}

\item{verbose}{Logical. Be chatty?}
}
\value{
DuckDb database connection
}
\description{
Connect to (and possibly create) DuckDB for storing results
}
