% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_etl.R
\name{dbGetBGC}
\alias{dbGetBGC}
\title{Get sitenos for preselected points in BGC/district}
\usage{
dbGetBGC(con, bgc, district = NULL, maxPoints = 150)
}
\arguments{
\item{con}{postgres DBI connection}

\item{bgc}{A character specifying BGC}

\item{district}{A character specifying district or NULL}

\item{maxPoints}{Integer - max number of points}
}
\value{
a vector of sitenumbers
}
\description{
Get sitenos for preselected points in BGC/district
}
\details{
Get siteno for BGC and District.
}
