% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_etl.R
\name{dbGetCCISS_novelty}
\alias{dbGetCCISS_novelty}
\title{Pull CCISS version 13 from a vector of SiteNo with novelty}
\usage{
dbGetCCISS_novelty(
  con,
  siteno,
  avg,
  modWeights,
  nov_cutoff = 5,
  cciss_table = "cciss_future_array",
  novelty_table = "cciss_novelty_array",
  cciss_observed = "cciss_current_nov",
  bgc_table = "bgc_attribution",
  bgc_lookup = "bgc_v13"
)
}
\arguments{
\item{con}{An active postgres DBI connection.}

\item{siteno}{A character vector of siteno.}

\item{avg}{A boolean.}

\item{modWeights}{A data table of gcm and rcp weights.}

\item{nov_cutoff}{Numeric between 1 and 8. Sigma value specifying novelty cutoff.}
}
\value{
A data.table containing CCISS information for each provided SiteNo.
}
\description{
Pull CCISS version 13 from a vector of SiteNo with novelty
}
\details{
Get CCISS for provided SiteNo.
}
