% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_etl.R
\name{dbGetCCISS_v13}
\alias{dbGetCCISS_v13}
\title{Pull CCISS version 13 from a vector of SiteNo}
\usage{
dbGetCCISS_v13(
  con,
  siteno,
  avg,
  modWeights,
  cciss_table = "cciss_novelty_array",
  cciss_observed = "cciss_current",
  bgc_table = "bgc_attribution",
  bgc_lookup = "bgc_v13"
)
}
\arguments{
\item{con}{An active postgres DBI connection.}

\item{siteno}{A character vector of siteno.}

\item{avg}{A boolean.}

\item{modWeights}{A data table of gcm and rcp weights.}
}
\value{
A data.table containing CCISS information for each provided SiteNo.
}
\description{
Pull CCISS version 13 from a vector of SiteNo
}
\details{
Get CCISS for provided SiteNo.
}
