% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_duckdb.R
\name{dbPopulate}
\alias{dbPopulate}
\title{Populate database with bgc point, edatopic and suitability tables}
\usage{
dbPopulate(dbCon, bgc_template, edatopes = c("B2", "C4", "D6"))
}
\arguments{
\item{dbCon}{Database connection}

\item{edatopes}{Character vector of edatopes}
}
\description{
Populate database with bgc point, edatopic and suitability tables
}
