% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climr-wrappers.R
\name{.getClimVars}
\alias{.getClimVars}
\title{Wrapper for \code{climr::downscale} that keeps extra columns in
table of point coordinates.}
\usage{
.getClimVars(
  coords,
  byCombo = FALSE,
  outFormat = "data.table",
  filename = tempfile(fileext = ".rds"),
  ...
)
}
\arguments{
\item{coords}{a \code{data.table}, or spatial points (\code{SpatVector} or \code{sf})
with point coordinates ("lon" = longitude, "lat" = latitude), elevation ("elev") and point IDs ("id").}

\item{byCombo}{logical. If TRUE, \code{climr::downscale} is iterated by
combinations of any arguments passed to \code{downscale} via  \code{...}
(e.g., \code{vars}, \code{gcms}, \code{ssps}).}

\item{outFormat}{character. Should outputs be in the form of a
\code{data.table} ("data.table"), list of \code{data.tables} ("list") or
written directly to disk ("disk")? If \code{byCombo == FALSE}, only
"data.table" and "disk" will work.}

\item{filename}{character. Passed to \code{saveRDS(..., file)} if
\code{outFormat == "disk"}. Defaults to \code{tempfile(fileext = ".csv")}.}

\item{...}{further arguments passed to \code{\link[climr:downscale]{climr::downscale()}}.}
}
\value{
climate normals returned as a \code{data.table} or list of \code{data.tables}.
If \code{outFormat == "disk"}, the file name(s) are returned instead.
}
\description{
Wrapper for \code{climr::downscale} that keeps extra columns in
table of point coordinates.
}
\details{
If \code{outFormat == "disk"} and \code{byCombo == TRUE}, \code{climr::downscale}
is iterated for combinations of GCMs, runs, periods and scenarios,
and each output \code{data.table} is saved to an \code{.rds} file with
\code{saveRDS(..., file = filename2, append = TRUE)}, where \code{basename(filename2)}
is the \code{basename(filename)} with appended values of the arguments that were iterated
through -- excluding the arguments "cache", "xyz", "nthread", "max_run",
"return_refperiod", "ppt_lr", "out_spatial", and "plot".
}
