% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climr-wrappers.R
\name{.renameUniqueCol}
\alias{.renameUniqueCol}
\title{Rename a column}
\usage{
.renameUniqueCol(x, otherCols)
}
\arguments{
\item{x}{a \code{data.table}.}

\item{otherCols}{character. Columns not to be renamed}
}
\value{
x with target column renamed as "value" and a
new column "var" containing the original name of the
target column
}
\description{
Rename a column
}
