% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_duckdb.R
\name{duckdb_table_exists}
\alias{duckdb_table_exists}
\title{Check in table exisits in current database}
\usage{
duckdb_table_exists(dbCon, table_name)
}
\arguments{
\item{table_name}{Character}

\item{dbcon}{duckdb connection}
}
\value{
Logical
}
\description{
Check in table exisits in current database
}
