% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objStorage.R
\name{dwnldFromObjSto}
\alias{dwnldFromObjSto}
\title{Download a file from Object Storage}
\usage{
dwnldFromObjSto(prefix, bucket, path)
}
\arguments{
\item{prefix}{passed to \code{\link[aws.s3:sync]{aws.s3::s3sync()}}. Must be a \emph{path}
to a remote folder. Use trailing slash (\verb{~/}).}

\item{bucket}{passed to \code{\link[aws.s3:sync]{aws.s3::s3sync()}}.}

\item{path}{passed to \code{\link[aws.s3:sync]{aws.s3::s3sync()}}. Must be an \emph{existing path}
to a local folder.}
}
\description{
Download a file from Object Storage
}
\details{
It is assumed that th object storage connection details and
credentials have already been set up as system environment
variables (notably "AWS_ACCESS_KEY_ID", "AWS_SECRET_ACCESS_KEY"
and "AWS_S3_ENDPOINT").
Note that all objects inside \code{path} will be synced to \code{path}
}
