% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_overlap.R
\name{edatopicOverlap_fast}
\alias{edatopicOverlap_fast}
\title{Pared down edatopic overlap used for CCISS spatial which ignores phases/variations}
\usage{
edatopicOverlap_fast(BGC, E1)
}
\arguments{
\item{BGC}{data.table of summarised BGC projections}

\item{E1}{Edatopic table}
}
\value{
data.table of site series predictions
}
\description{
Pared down edatopic overlap used for CCISS spatial which ignores phases/variations
}
