% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_analysis.R
\name{ensemble_novelty}
\alias{ensemble_novelty}
\title{Calculate mean novelty across runs}
\usage{
ensemble_novelty(con, table_name = "ensemble_novelty")
}
\arguments{
\item{con}{duckdb database connection}

\item{table_name}{name of table to create with mean novelty values}
}
\description{
Calculate mean novelty across runs
}
