% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_duckdb.R
\name{ensemble_predictions}
\alias{ensemble_predictions}
\title{Create ensemble subzone or zone winner}
\usage{
ensemble_predictions(
  dbCon,
  ssp_use = c("ssp126", "ssp245", "ssp370"),
  ssp_w = c(0.8, 1, 0.8),
  by_zone = FALSE,
  table_name = "ensemble_preds"
)
}
\arguments{
\item{ssp_use}{Character. List of ssps to use. Default \code{c("ssp126", "ssp245", "ssp370")}}

\item{ssp_w}{Numeric vector. Weights for each ssp in \code{ssp_use}}

\item{by_zone}{Logical. Calculate winner by zone? Default \code{FALSE}.}

\item{table_name}{Name of table to create in database. Table must not already exist.}
}
\value{
NULL. Table is written to duckdb connection
}
\description{
Create ensemble subzone or zone winner
}
