% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-utils.R
\name{GDALcrop}
\alias{GDALcrop}
\title{Crop on disk using \code{gdalUtilities::ogr2ogr}}
\usage{
GDALcrop(filename, filename2, destinationPath = NULL, studyArea = NULL)
}
\arguments{
\item{filename}{character. File name (only) to original spatial file.}

\item{filename2}{character. File name (only) to use when saving cropped spatial file.}

\item{destinationPath}{folder path to \code{filename} and where \code{filename2} will be saved.}

\item{studyArea}{a SpatVector, or NULL (default), from which the extent to crop
\code{filename} to will be extracted. If NULL, no cropping/saving happens and \code{filename} is
simply loaded.}
}
\value{
a cropped SpatVector that corresponds to \code{filename2}.
}
\description{
Crop on disk using \code{gdalUtilities::ogr2ogr}
}
