% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climr-wrappers.R
\name{getClimate}
\alias{getClimate}
\title{Prepares coordinates and obtains climate normals
using \code{climr_downscale}}
\usage{
getClimate(coords, ...)
}
\arguments{
\item{coords}{a \code{data.table}, or spatial points (\code{SpatVector} or \code{sf})
with point coordinates ("lon" = longitude, "lat" = latitude), elevation ("elev") and point IDs ("id").}

\item{...}{further arguments passed to \code{\link[=climr_downscale]{climr_downscale()}}.}
}
\value{
climate normals as a \code{data.table}
}
\description{
Prepares coordinates and obtains climate normals
using \code{climr_downscale}
}
\details{
If \code{bgcs} is provided, the BGC field will be appended to
the output \code{data.table}.
}
\seealso{
\code{\link[=climr_downscale]{climr_downscale()}}
}
