% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logVars.R
\name{logVars}
\alias{logVars}
\title{Log-transform climate variables}
\usage{
logVars(
  dat,
  elements = c("AHM", "DD", "Eref", "FFP", "NFFD", "PAS", "PPT", "SHM", "CMD"),
  base = exp(1),
  add.fields = FALSE,
  zero_adjust = FALSE
)
}
\arguments{
\item{dat}{a \code{data.table} with columns of climate variables corresponding to
the selected climate \code{elements}.}

\item{elements}{character. Climate elements to search for in \code{dat}.}

\item{base}{numeric. Log base.}

\item{add.fields}{logical. If \code{TRUE}, the new logged variables are added to \code{dat}
(TRUE). Otherwise, original column values are replaced with the logs (FALSE).}

\item{zero_adjust}{logical. If \code{TRUE} adjusts zeroes in raw data as:
\eqn{base^{\log_base{x_min} - 1}}.
where \eqn{x_min} is the minimum non-zero, non-NA value.}
}
\value{
\code{data.table}
}
\description{
Log-transform climate variables
}
\details{
All column names that partially match strings in \code{elements} will be
log-transformed.
}
