% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_spatial.R
\name{make_bgc_raster}
\alias{make_bgc_raster}
\title{Create raster and id table of BGCs based on vector data}
\usage{
make_bgc_raster(raster_template, bgcs)
}
\arguments{
\item{raster_template}{SpatRaster.}

\item{bgcs}{SpatVector or sf object of bgc bondaries with column "BGC" specifying name.}
}
\value{
List containinng resultant SpatRaster and data.table of ids.
}
\description{
Create raster and id table of BGCs based on vector data
}
