% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_spatial.R
\name{make_bgc_template}
\alias{make_bgc_template}
\title{Create raster and id table of BGCs based on vector data}
\usage{
make_bgc_template(xyz, bgcs, res = 8e-04)
}
\arguments{
\item{xyz}{SpatRaster or data.table of points (must have lat, lon, id).}

\item{bgcs}{SpatVector or sf object of bgc bondaries with column "BGC" specifying name.}
}
\value{
List containing resultant SpatRaster and data.table of ids (or single data.table if input is a data.table).
}
\description{
Create raster and id table of BGCs based on vector data
}
