% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-utils.R
\name{makeGapExtents}
\alias{makeGapExtents}
\title{Make generate extents for gaps used in hold-out samples
for model cross-validation.}
\usage{
makeGapExtents(studyarea = ext(c(-123, -118, 49, 52)), ngaps = 5L)
}
\arguments{
\item{studyarea}{\code{SpatExtent} of study area where gaps should be generated.
Defaults to an area in Southern BC.}

\item{ngaps}{integer. Number of gaps wanted..}
}
\value{
a \code{list} of extents of length \code{ngaps}.
}
\description{
Make generate extents for gaps used in hold-out samples
for model cross-validation.
}
