% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-utils.R
\name{makePointCoords}
\alias{makePointCoords}
\title{Create a set of training points with associated
elevation and BGC values.}
\usage{
makePointCoords(bgc_poly, elev, gridSize = 2000, crs = "EPSG:4326")
}
\arguments{
\item{bgc_poly}{an \code{sf} object (or one cohercible to an \code{sf} object) of BGC polygons.}

\item{elev}{a \code{SpatRaster} or \code{RasterLayer} of elevation covering the extent of \code{bgc_poly}.}

\item{gridSize}{numeric. Distance in m between points.}

\item{crs}{passed to \code{\link[sf:st_as_sf]{sf::st_as_sf()}}}
}
\value{
a \code{data.table} of point coordinates with associated IDs,
elevation and BGCs.
}
\description{
Create a set of training points with associated
elevation and BGC values.
}
\details{
Points are sampled regularly from a grid with cell size
defined by \code{gridSize} that covers \code{bgc_poly}.
}
