% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_spatial.R
\name{mapped_suit}
\alias{mapped_suit}
\title{Function to create rasters of historic (mapped) suitability by species and edatopic position}
\usage{
mapped_suit(species, edatopes, bgc_template_list, base_folder)
}
\arguments{
\item{species}{Character. Vector of species codes to map}

\item{edatopes}{Character. Vector of edatopes (e.g., "C4")}

\item{bgc_template_list}{List containing SpatRaster of BGCs and id table. Usually created using \code{make_bgc_template}}

\item{base_folder}{Base folder to write results to.}
}
\value{
NULL. Writes tifs to "base_folder/historic_suit"
}
\description{
Function to create rasters of historic (mapped) suitability by species and edatopic position
}
