% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_qml.R
\name{parse_qml}
\alias{parse_qml}
\title{Parse QML stylesheet to table of hex values}
\usage{
parse_qml(qml_path)
}
\arguments{
\item{qml_path}{Character. Path of qml stylesheet}
}
\value{
A data.table containing columns classification and colour
}
\description{
Parse QML stylesheet to table of hex values
}
\details{
Parse stylesheet
}
