% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_analysis.R
\name{plot_2panel}
\alias{plot_2panel}
\title{Create C.R. Mahony's two-panel map plot of historic suitability, suitability change, and a boxplot of change by zone}
\usage{
plot_2panel(
  spp,
  edatope,
  period,
  bgc_template,
  outline,
  save_png = TRUE,
  base_path
)
}
\arguments{
\item{spp}{Character. Species code to create plot for}

\item{edatope}{Character. Edatope code to create plot for (e.g., "C4")}

\item{period}{Character. Period to create plot for.}

\item{bgc_template}{List containing SpatRaster of BGCs and id table. Usually created using \code{make_bgc_template}}

\item{outline}{SpatVector of aoi boundary.}

\item{save_png}{Logical. Save plot to png? Default \code{TRUE}. If \code{FALSE}, creates plot on default plotting device.}

\item{base_folder}{Base folder to read results from.}
}
\value{
NULL. Creates plot
}
\description{
Note that currently, this function only works correctly if the preceeding analysis has been done using an Albers grid.
}
