% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_analysis.R
\name{plot_alluvial}
\alias{plot_alluvial}
\title{Create alluvial/stacked bar plot of projected area by zone/subzone}
\usage{
plot_alluvial(dat, spp, edatope, by_zone = T, cellarea = 4)
}
\arguments{
\item{dat}{data.table of species area data from 'alluvial_area' table}

\item{spp}{Character. Single species to use for plot}

\item{edatope}{Character. Single edatopic position for plot (e.g., "C4")}

\item{by_zone}{Logical. Plot by zone or subzone? Defaults to \code{TRUE}}
}
\value{
NULL. Creates plot
}
\description{
Create alluvial/stacked bar plot of projected area by zone/subzone
}
