% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_analysis.R
\name{plot_bgc}
\alias{plot_bgc}
\title{Save raster of predicted BGC map}
\usage{
plot_bgc(
  con,
  period,
  plot_ensemble,
  raster_template,
  gcm = NULL,
  ssp = NULL,
  run = NULL,
  by_zone = FALSE,
  save_location = "bgc_rasters",
  raw = FALSE,
  plot = FALSE
)
}
\arguments{
\item{con}{duckdb connection}

\item{period}{Character. Period to plot for.}

\item{plot_ensemble}{Logical. Plot vote winner over ensemble of runs?}

\item{raster_template}{SpatRaster to use as template for plotting.}

\item{gcm, ssp, run}{Character. Only used if \code{plot_ensemble = FALSE}}

\item{by_zone}{Logical. Plot by zone or subzone variant? Default \code{FALSE}}

\item{save_location}{Character. Path to save raster in. Will be created if it doesn't exist.}

\item{raw}{Logical. Save raw values, or colourised rgb rasters? Default \code{FALSE}}

\item{plot}{Logical. Create plot as well as saving raster? Default \code{FALSE}}
}
\description{
Save raster of predicted BGC map
}
