% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_analysis.R
\name{plot_bgc_mapped}
\alias{plot_bgc_mapped}
\title{Save raster of actual BGC map}
\usage{
plot_bgc_mapped(
  raster_template,
  by_zone = FALSE,
  save_location = "bgc_rasters",
  plot = FALSE
)
}
\arguments{
\item{raster_template}{List returned from \code{create_bgc_template()}}

\item{by_zone}{Logical. Plot by zone or subzone variant? Default \code{FALSE}}

\item{save_location}{Character. Path to save raster in. Will be created if it doesn't exist.}

\item{plot}{Logical. Create plot as well as saving raster? Default \code{FALSE}}
}
\description{
Save raster of actual BGC map
}
