% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_analysis.R
\name{plot_novelty}
\alias{plot_novelty}
\title{Plot novelty map with values from database}
\usage{
plot_novelty(
  con,
  raster_template,
  period,
  ensemble = TRUE,
  gcm = NULL,
  ssp = NULL,
  run = NULL,
  save_location = "novelty_rasters"
)
}
\arguments{
\item{con}{duckdb database connection}

\item{raster_template}{SpatRaster template to use for plotting}

\item{period}{Character. Which period to plot for?}

\item{ensemble}{Logical. Plot ensemble novelty (default) or raw novelty for a specified run?}
}
\description{
Plot novelty map with values from database
}
