% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_analysis.R
\name{plot_predicted_suitability}
\alias{plot_predicted_suitability}
\title{Save raster of predicted suitability for use in e.g. CCISS Spatial}
\usage{
plot_predicted_suitability(
  con,
  spp,
  period,
  edatope,
  raster_template,
  save_location = "suit_rasters",
  raw = FALSE,
  plot = FALSE
)
}
\arguments{
\item{con}{duckdb connection}

\item{spp}{Species to plot}

\item{period}{Character. Period to plot for.}

\item{edatope}{Character. Edatope position (e.g. C4)}

\item{raster_template}{SpatRaster to use as template for plotting.}

\item{save_location}{Character. Path to save raster in. Will be created if it doesn't exist.}

\item{raw}{Logical. Save raw values, or colourised rgb rasters? Default \code{FALSE}}

\item{plot}{Logical. Create plot as well as saving raster? Default \code{FALSE}}
}
\description{
Save raster of predicted suitability for use in e.g. CCISS Spatial
}
