% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_analysis.R
\name{plot_reference_suitability}
\alias{plot_reference_suitability}
\title{Plot map of reference suitability for given species and edatope}
\usage{
plot_reference_suitability(
  con,
  raster_template,
  outline,
  spp,
  edatope = "C4",
  save_png = FALSE
)
}
\arguments{
\item{con}{duckdb database connection}

\item{raster_template}{SpatRaster to use as template for plotting.}

\item{outline}{SpatVector of aoi boundary.}

\item{spp}{Character. Species code to create plot for}

\item{edatope}{Character. Edatope code to create plot for (e.g., "C4")}

\item{save_png}{Logical. Save plot to png? Default \code{FALSE}. If \code{FALSE}, creates plot on default plotting device.}
}
\value{
NULL. Creates plot
}
\description{
Note that currently, this function only works correctly if the preceeding analysis has been done using an Albers grid.
}
