% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_analysis.R
\name{plot_SuitabilityChangeMap}
\alias{plot_SuitabilityChangeMap}
\title{Create C.R. Mahony's two-panel map plot of historic suitability, suitability change, and a boxplot of change by zone}
\usage{
plot_SuitabilityChangeMap(
  dbCon,
  bgc_template,
  outline,
  spp = "Fd",
  edatope = "C4",
  period = "2041_2060",
  three_panel = FALSE,
  save_png = TRUE,
  panel_labels = TRUE
)
}
\arguments{
\item{dbCon}{duckdb database connection}

\item{bgc_template}{List containing SpatRaster of BGCs and id table. Usually created using \code{make_bgc_template}}

\item{outline}{SpatVector of aoi boundary.}

\item{spp}{Character. Species code to create plot for}

\item{edatope}{Character. Edatope code to create plot for (e.g., "C4")}

\item{period}{Character. Period to create plot for.}

\item{save_png}{Logical. Save plot to png? Default \code{TRUE}. If \code{FALSE}, creates plot on default plotting device.}

\item{panel_labels}{Logical. Add a manuscript-style label to each panel of the plot.}
}
\value{
NULL. Creates plot
}
\description{
Note that currently, this function only works correctly if the preceeding analysis has been done using an Albers grid.
}
