% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_spatial.R
\name{predict_bgc}
\alias{predict_bgc}
\title{Creates BGC predictions from RF model}
\usage{
predict_bgc(
  dbCon,
  xyz,
  BGCmodel,
  vars_needed,
  gcms_use,
  periods_use,
  ssp_use = c("ssp126", "ssp245", "ssp370"),
  obs_2001_2020 = FALSE,
  max_runs_use = 0L,
  start_tile = 1
)
}
\arguments{
\item{dbCon}{Database connection}

\item{xyz}{SpatRaster or data.table.}

\item{BGCmodel}{Ranger random forest model of BGCs}

\item{vars_needed}{Character. List of variables required for model}

\item{gcms_use}{Character. List of gcms used in summarised predictions}

\item{periods_use}{Character. List of time periods to create predictions for}

\item{ssp_use}{Character. List of ssps to use. Default \code{c("ssp126", "ssp245", "ssp370")}}

\item{obs_2001_2020}{Logical. Also predict for 2001 - 2020 observed climate? Default \code{FALSE}}

\item{max_runs_use}{Integer. Number of individual runs for each GCM. Default \code{0} (only use ensembleMean)}
}
\value{
NULL. Results are written to csv files in base_folder/bgc_data
}
\description{
Creates BGC predictions from RF model
}
