% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_spatial.R
\name{ref_suit}
\alias{ref_suit}
\title{Function to create create table of reference species suitability by BGC for given edatopic positions}
\usage{
ref_suit(dbCon, species, edatopes)
}
\arguments{
\item{dbCon}{duckdb database connection}

\item{species}{Character. Vector of species codes to map}

\item{edatopes}{Character. Vector of edatopes (e.g., "C4")}
}
\value{
Data.table of species x edatope x bgc suitabilities
}
\description{
Function to create create table of reference species suitability by BGC for given edatopic positions
}
