% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeOutlier.R, R/stats.R
\name{removeOutlier}
\alias{removeOutlier}
\title{Remove outliers from data}
\usage{
removeOutlier(dat, alpha, vars)

removeOutlier(dat, alpha, vars)
}
\arguments{
\item{dat}{a \code{data.table} target data to "clean"}

\item{alpha}{numeric. The alpha value used to determine the cutoff for outliers.}

\item{vars}{character. Names of columns from which outliers should be excluded.}

\item{IDvars}{character. Names of columns from which outliers should be excluded.}
}
\value{
A \code{data.table} containing the non-outlying observations

\code{data.table}
}
\description{
Remove outliers from data

Remove outliers from data
}
\details{
TODO. Parallelizes computations internally with \code{foreach}.

TODO. Parallelizes computations internally with \code{foreach}.
}
\seealso{
\code{\link[foreach:foreach]{foreach::foreach()}}

\code{\link[foreach:foreach]{foreach::foreach()}}
}
