% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_spatial.R
\name{siteseries_preds}
\alias{siteseries_preds}
\title{Create siteseries predictions from summarised BGC predictions}
\usage{
siteseries_preds(
  edatopes = c("B2", "C4", "D6"),
  obs = F,
  bgc_mapped,
  base_folder = "spatial"
)
}
\arguments{
\item{edatopes}{Character. Vector of desired edatopic positions. Default is "B2", "C4", "D6" (poor, mesic, rich).}

\item{obs}{Logical. Do prediction for observed period? Default FALSE}

\item{bgc_mapped}{List containing SpatRaster of BGCs and id table, or data.table (must have columns \code{cell},\code{BGC},\code{bgc_id}). Usually created using \code{make_bgc_raster}}

\item{base_folder}{Base folder to write results to.}
}
\value{
NULL. Writes results to csvs in base_folder/ss_preds
}
\description{
Create siteseries predictions from summarised BGC predictions
}
