% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_duckdb.R
\name{spp_loss_gain}
\alias{spp_loss_gain}
\title{Create table of species relative suitabile area}
\usage{
spp_loss_gain(
  con,
  edatope = "C4",
  fut_wt = c(0, 0, 0, 0.5, 0.5, 0),
  curr_wt = c(0.5, 0.5, 0, 0, 0, 0),
  periods = c("Curr", "2001_2020", "2021_2040", "2041_2060", "2061_2080", "2081_2100"),
  BGCxDistrict = FALSE
)
}
\arguments{
\item{con}{duckdb database connection}

\item{edatope}{Character. Edatopic position to use. Default \code{C4} (zonal)}

\item{fut_wt}{Numeric vector of length 5, corresponding to weight of each period for future period calculation. Default \code{c(0,0,0,0.5,0.5,0)}, which is average of 2041-2060 and 2061-2080 periods.}

\item{curr_wt}{Numeric vector of length 5, corresponding to weight of each period for current period calculation. Default \code{c(0.5,0.5,0,0,0,0)}, which is average of refperiod and 2001-2020 periods.}

\item{periods}{Character vector of length 5 with names of periods. Must match names in table \code{cciss_res}. Default \code{c("Curr","2001_2020","2021_2040","2041_2060","2061_2080","2081_2100")}}

\item{BGCxDistrict}{Logical. Should summary be by BGC (default) or by BGC and natural resource district?}
}
\value{
data.table containing loss and gain proportion by Spp and BGC
}
\description{
Create table of species relative suitabile area
}
