% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_duckdb.R
\name{spp_persist_expand}
\alias{spp_persist_expand}
\title{Create table of species persistance/expansion from raw BGC projections}
\usage{
spp_persist_expand(dbCon, spp_list, fractional = TRUE)
}
\arguments{
\item{dbCon}{duckdb database connection}

\item{fractional}{Logical. Use fractional (suitability based) values for calculations?}

\item{species}{Character. Vector of species codes to use.}
}
\value{
data.table containing persistance and expansion for each model/period/scenario/species
}
\description{
Create table of species persistance/expansion from raw BGC projections
}
\details{
The calculation of species persistance can take a while - about 15 min for a 2km grid of BC. This function uses smart caching to try and speed up workflows. The first time the function is run, it caches compute-heavy tables that can be reused, as well as caching final results. Only species that have not yet been calculated will be processed.
}
