% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_duckdb.R
\name{spp_persist_expand_region}
\alias{spp_persist_expand_region}
\title{Calculate species persistance and expansion by region}
\usage{
spp_persist_expand_region(
  con,
  spp_list,
  region_table,
  region_name,
  fractional = TRUE,
  table_name = "spp_per_exp_region"
)
}
\arguments{
\item{con}{duckdb connection}

\item{spp_list}{character vector of species to use.}

\item{region_table}{character. Name of table with regions and corresponding cellnum}

\item{region_name}{character. Name of column in region table to group by}

\item{fractional}{logical. Summarise by fractional or binary suitability? Default \code{TRUE}.}

\item{table_name}{table name to create/append to in database. Default is "spp_per_exp_region"}
}
\value{
NULL. Writes table to database.
}
\description{
Calculate species persistance and expansion by region
}
