% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_duckdb.R
\name{spp_suit_area}
\alias{spp_suit_area}
\title{Create table of species relative suitabile area}
\usage{
spp_suit_area(dbCon, spp_list, fractional = TRUE)
}
\arguments{
\item{dbCon}{duckdb database connection}

\item{spp_list}{Character vector of species to analyse}

\item{fractional}{Logical. Use fractional (suitability based) values for calculations?}
}
\value{
data.table containing relative area for each run/model/period/scenario/species
}
\description{
Create table of species relative suitabile area
}
