% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-utils.R
\name{subsetByExtent}
\alias{subsetByExtent}
\title{Subset a group of points by spatial extent}
\usage{
subsetByExtent(coords, cropExt = ext(c(-123, -118, 49, 52)), crs = "EPSG:4326")
}
\arguments{
\item{coords}{\code{data.table} of point coordinates with columns "x" (longitude)
and "y" (latitude) (and any additional columns), a \code{SpatVector} or object
cohersible to \code{SpatVector}.}

\item{cropExt}{\code{SpatExtent} or \code{SpatVector} (whose extent will be used)
to subset the data to. Defaults to the \code{SpatExtent} of an area in Southern BC.}

\item{crs}{passed to \code{\link[terra:vect]{terra::vect()}} to coerce coords to a \code{SpatVector}
if it is not one already.}
}
\value{
a subset \code{coords} object.
}
\description{
Subset a group of points by spatial extent
}
