% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_duckdb.R
\name{summarise_preds}
\alias{summarise_preds}
\title{Summarise raw BGC predictions in database}
\usage{
summarise_preds(
  dbCon,
  ssp_use = c("ssp126", "ssp245", "ssp370"),
  ssp_w = c(0.8, 1, 0.8)
)
}
\arguments{
\item{ssp_use}{Character. List of ssps to use. Default \code{c("ssp126", "ssp245", "ssp370")}}

\item{ssp_w}{Numeric vector. Weights for each ssp in \code{ssp_use}}

\item{base_folder}{Character. Name of base folder to write results to.}
}
\value{
NULL. Results are written to csv files in base_folder/bgc_data
}
\description{
Summarise raw BGC predictions in database
}
