% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_spatial.R
\name{summary_preds_gcm}
\alias{summary_preds_gcm}
\title{Create summarised BGC predictions from RF model}
\usage{
summary_preds_gcm(
  xyz,
  BGCmodel,
  vars_needed,
  gcms_use,
  periods_use,
  ssp_use = c("ssp126", "ssp245", "ssp370"),
  ssp_w = c(0.8, 1, 0.8),
  base_folder = "spatial",
  start_tile = 1
)
}
\arguments{
\item{xyz}{SpatRaster or data.table.}

\item{BGCmodel}{Ranger random forest model of BGCs}

\item{vars_needed}{Character. List of variables required for model}

\item{gcms_use}{Character. List of gcms used in summarised predictions}

\item{periods_use}{Character. List of time periods to create predictions for}

\item{ssp_use}{Character. List of ssps to use. Default \code{c("ssp126", "ssp245", "ssp370")}}

\item{ssp_w}{Numeric vector. Weights for each ssp in \code{ssp_use}}

\item{base_folder}{Character. Name of base folder to write results to.}
}
\value{
NULL. Results are written to csv files in base_folder/bgc_data
}
\description{
Create summarised BGC predictions from RF model
}
