% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVars.R
\name{addVars}
\alias{addVars}
\title{Add extra climate variables}
\usage{
addVars(dat)
}
\arguments{
\item{dat}{a \code{data.table} with columns:
\itemize{
\item PPT05, PPT06, PPT07, PPT08, PPT09 (May, ..., September precip),
PPT_at (autumn precip), PPT_wt (winter precip)
\item CMD07 (July climate moisture deficit), CMD (annual CMD)
\item DD_0_at (autumn degree-days below 0 deg), DD_0_wt (winter degree-days below 0 deg)
}}
}
\value{
\code{dat} with all of the above added climate variables.
}
\description{
Add extra climate variables
}
\details{
This function calculates more climate variables derived from those
output by \code{climr_downscale}. Presently it adds the following:
\itemize{
\item May to June precip.: \eqn{PPT_MJ = PPT05 + PPT06}
\item July to September precip \eqn{PPT_JAS = PPT07 + PPT08 + PPT09}
\item Precipitation during vegetation dormant period: \eqn{PPT.dormant = PPT_at + PPT_wt}
\item CMD deficit \eqn{CMD.def = 500 - PPT.dormant} (bounded to 0)
\item \eqn{CMDMax = CMD07}
\item \eqn{CMD.total = CMD.def + CMD}
\item \eqn{DD_delayed = ((DD_0_at + DD_0_wt)*0.0238) - 1.8386} bounded to 0)
}
}
