% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_data.R
\docType{data}
\name{bccciss-data}
\alias{bccciss-data}
\alias{BGCRegions}
\alias{cfrg_rules}
\alias{covMat}
\alias{E1}
\alias{E1_Phase}
\alias{F1}
\alias{footnotes}
\alias{models_info}
\alias{N1}
\alias{R1}
\alias{SIBEC}
\alias{silvics_mature}
\alias{silvics_regen}
\alias{silvics_resist}
\alias{silvics_tol}
\alias{SS}
\alias{stocking_height}
\alias{stocking_info}
\alias{stocking_standards}
\alias{subzones_colours_ref}
\alias{T1}
\alias{TreeCols}
\alias{V1}
\alias{zones_colours_ref}
\title{Data to be included in ccissr package}
\description{
Data to be included in ccissr package
}
\keyword{data}
