% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_analysis.R
\name{bgc_bubbleplot}
\alias{bgc_bubbleplot}
\title{Create C.R. Mahony's bubbleplot of bgc expansion/persistance}
\usage{
bgc_bubbleplot(
  persist_expand,
  period = "2041_2060",
  scenario = "ssp245",
  xlab = "Climate analog within historical range",
  ylab = "Climate analog beyond historical range",
  unit.focal = NULL,
  focal.color = "lightskyblue2",
  xlim = c(0, 1.1),
  ylim = c(-6, 4),
  xlabels = TRUE,
  ylabels = TRUE,
  mar = c(3, 4, 0.1, 0.1),
  plt = NULL
)
}
\arguments{
\item{persist_expand}{data.table. Usually created with \code{bgc_persist_expand}}

\item{period}{Character. Period to create plot for.}

\item{scenario}{Character. ssp to use for plot.}

\item{xlab}{Character. x axis title.}

\item{ylab}{Character. y axis title.}

\item{xlabels}{logical. x axis labels.}

\item{ylabels}{logical. y axis labels.}

\item{mar}{numeric. plot margins.}

\item{plt}{numeric. par(plt) parameters for overlaying into an existing plot window.}
}
\value{
NULL. Creates plot
}
\description{
Create C.R. Mahony's bubbleplot of bgc expansion/persistance
}
