% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_analysis.R
\name{bgc_bubbleplot}
\alias{bgc_bubbleplot}
\title{Create C.R. Mahony's bubbleplot of bgc expansion/persistance}
\usage{
bgc_bubbleplot(persist_expand, period, scenario)
}
\arguments{
\item{persist_expand}{data.table. Usually created with \code{bgc_persist_expand}}

\item{period}{Character. Period to create plot for.}

\item{scenario}{Character. ssp to use for plot.}
}
\value{
NULL. Creates plot
}
\description{
Create C.R. Mahony's bubbleplot of bgc expansion/persistance
}
