% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_duckdb.R
\name{bgc_persist_expand}
\alias{bgc_persist_expand}
\title{Create table of BGC subzone/zone persistance/expansion from raw BGC projections}
\usage{
bgc_persist_expand(dbCon, by_zone = TRUE)
}
\arguments{
\item{dbCon}{database connection to duckdb}

\item{by_zone}{Logical. Summarise by Zone, or subzone variant? Default \code{TRUE}}
}
\value{
data.table containing persistance and expansion for each model/period/scenario/bgc
}
\description{
Create table of BGC subzone/zone persistance/expansion from raw BGC projections
}
\details{
This function will cache the resulting table in the database for faster future use.
}
