% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_spatial.R
\name{cciss_suitability}
\alias{cciss_suitability}
\title{Create projected suitability values from site series predictions.}
\usage{
cciss_suitability(dbCon, species, obs = FALSE, tile_size = 4000)
}
\arguments{
\item{species}{Character vector. Species codes to create projections for.}

\item{tile_size}{Integer. Number of sites to process at once. May need to decrease if memory is limited. Default 4000}

\item{edatopes}{Character vector. List of edatopes for projections. Must be a subset of the edatopes run in \code{siteseries_preds}}

\item{periods}{Character. Periods to run projections for. Default is \code{list_gcm_periods()}}

\item{base_folder}{Base folder to write results to.}
}
\value{
NULL. Writes results to csvs in base_folder/cciss_suit
}
\description{
Create projected suitability values from site series predictions.
}
