% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climr-wrappers.R
\name{.getClimVars}
\alias{.getClimVars}
\title{Wrapper for \code{climr_downscale} that keeps extra columns in
table of point coordinates.}
\usage{
.getClimVars(coords, ...)
}
\arguments{
\item{coords}{a \code{data.table}, or spatial points (\code{SpatVector} or \code{sf})
with point coordinates ("lon" = longitude, "lat" = latitude), elevation ("elev") and point IDs ("id").}

\item{...}{further arguments passed to \code{\link[=climr_downscale]{climr_downscale()}}.}
}
\value{
climate normals as a \code{data.table}.
}
\description{
Wrapper for \code{climr_downscale} that keeps extra columns in
table of point coordinates.
}
