% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_analysis.R
\name{plot_spparea}
\alias{plot_spparea}
\title{Create alluvial/stacked bar plot of projected area by zone/subzone}
\usage{
plot_spparea(
  dbCon,
  spp,
  edatope,
  fractional,
  by_zone = TRUE,
  save_png = TRUE,
  width = 6,
  height = 7,
  res = 300
)
}
\arguments{
\item{spp}{Character. Single species to use for plot}

\item{edatope}{Character. Single edatopic position for plot (e.g., "C4")}

\item{fractional}{Logical. Use fractional (suitability based) values for calculations?}

\item{by_zone}{Logical. Plot by zone or subzone? Defaults to \code{TRUE}}

\item{save_png}{Logical. Save plot to png? Default \code{TRUE}. If \code{FALSE}, creates plot on default plotting device.}

\item{bgc_template}{List containing SpatRaster of BGCs and id table, or data.table (must have columns \code{cell},\code{BGC}). Usually created using \code{make_bgc_template}}

\item{base_folder}{Base folder to read results from.}
}
\value{
NULL. Creates plot
}
\description{
Create alluvial/stacked bar plot of projected area by zone/subzone
}
