% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_spatial.R
\name{siteseries_preds}
\alias{siteseries_preds}
\title{Create siteseries predictions from summarised BGC predictions}
\usage{
siteseries_preds(dbCon, obs = FALSE)
}
\arguments{
\item{dbCon}{Database connection to duckdb}
}
\description{
Create siteseries predictions from summarised BGC predictions
}
