% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_analysis.R
\name{spp_bubbleplot}
\alias{spp_bubbleplot}
\title{Create C.R. Mahony's bubbleplot of species expansion/persistance}
\usage{
spp_bubbleplot(
  persist_expand,
  period = "2041_2060",
  scenario = "ssp245",
  species = "auto",
  species.focal = NULL,
  focal.color = "lightskyblue2",
  edatope = "C4",
  by = "species",
  xlab = "Persistence within historically suitable range",
  ylab = "Expansion beyond historically suitable range",
  xlim = c(0, 1.15),
  ylim = c(-6, 4),
  xlabels = TRUE,
  ylabels = TRUE,
  mar = c(3, 4, 0.1, 0.1)
)
}
\arguments{
\item{persist_expand}{data.table. Usually created with \code{spp_persist_expand}}

\item{period}{Character. Period to create plot for. Must be one of climr::list_gcm_periods().}

\item{scenario}{Character. ssp to use for plot. Must be one of climr::list_spps().}

\item{species}{Character vector of species or "auto" to display all available species. Default "auto".}

\item{species.focal}{Character A single species for which to display time evolution and ensemble variation. Must be an element of the vector defined in the 'species' parameter. Default NULL.}

\item{focal.color}{Character A color for the focal species data points.}

\item{edatope}{Character of desired edatope to use for multiple species plot, or vector of edatopes for single species.}

\item{by}{Character. Either "species", to display multiple species for one edatope, or "edatopes", to display multiple edatopes for one species.}

\item{xlab}{Character. x axis title.}

\item{ylab}{Character. y axis title.}

\item{xlabels}{logical. x axis labels.}

\item{ylabels}{logical. y axis labels.}

\item{mar}{numeric. plot margins.}
}
\value{
NULL. Creates plot
}
\description{
Create C.R. Mahony's bubbleplot of species expansion/persistance
}
