% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_analysis.R
\name{spp_bubbleplot}
\alias{spp_bubbleplot}
\title{Create C.R. Mahony's bubbleplot of species expansion/persistance}
\usage{
spp_bubbleplot(
  persist_expand,
  period,
  scenario,
  species = "auto",
  edatope,
  by = "species"
)
}
\arguments{
\item{persist_expand}{data.table. Usually created with \code{spp_persist_expand}}

\item{period}{Character. Period to create plot for.}

\item{scenario}{Character. ssp to use for plot.}

\item{species}{Character vector of species or "auto" to display all available species. Default "auto".}

\item{edatope}{Character of desired edatope to use for multiple species plot, or vector of edatopes for single species.}

\item{by}{Character. Either "species", or "edatopes". If}
}
\value{
NULL. Creates plot
}
\description{
Create C.R. Mahony's bubbleplot of species expansion/persistance
}
