% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_analysis.R
\name{spp_persist_expand}
\alias{spp_persist_expand}
\title{Create table of species persistance/expansion from raw BGC projections}
\usage{
spp_persist_expand(dbCon, spp_list, fractional = TRUE)
}
\arguments{
\item{fractional}{Logical. Use fractional (suitability based) values for calculations?}

\item{edatopes}{Character. Vector of desired edatopic positions.}

\item{bgc_mapped}{List containing SpatRaster of BGCs and id table, or data.table (must have columns \code{cell},\code{BGC}). Usually created using \code{make_bgc_template}}

\item{species}{Character. Either "auto", or a vector of species codes to use. If "auto", all species with historically suitable range will be used.}

\item{periods}{Character. Either "auto" or vector of period names. If "auto", all periods present in bgc projections will be used.}

\item{base_folder}{Base folder to read results from.}
}
\value{
data.table containing persistance and expansion for each model/period/scenario/species
}
\description{
Create table of species persistance/expansion from raw BGC projections
}
