% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cciss_analysis.R
\name{spp_suit_area}
\alias{spp_suit_area}
\title{Create table of species relative suitabile area}
\usage{
spp_suit_area(edatopes, fractional, spp_list, bgc_mapped, base_folder)
}
\arguments{
\item{edatopes}{Character. Vector of desired edatopic positions.}

\item{fractional}{Logical. Use fractional (suitability based) values for calculations?}

\item{spp_list}{Character vector of species to analyse}

\item{bgc_mapped}{List containing SpatRaster of BGCs and id table, or data.table (must have columns \code{cell},\code{BGC}). Usually created using \code{make_bgc_template}}

\item{base_folder}{Base folder to read results from.}
}
\value{
data.table containing relative area for each run/model/period/scenario/species
}
\description{
Create table of species relative suitabile area
}
